add "\0" to send,  create printScreenHandler
